TST: Move slow tests to a new file for test_associations_sdp_pools to see if it helps with run time
