+++ .ruff.toml
@@ -126,6 +126,10 @@ ban-relative-imports = "all"  # Disallow all relative imports.
     "SLF", "ARG", "PTH",
     "B003", # Assigning to os.environ does not clear the environment
 ]
+"jwst/regtest/assoc_rt_helpers.py" = [
+    "SLF",  "ARG", "PTH",
+    "N813", # Camelcase imported as lowercase
+]
 "jwst/regtest/regtestdata.py" = [
     "SLF", "ARG", "PTH",
     "A002", # builtin variable shadowing
