+++ jwst/regtest/test_associations_sdp_pools.py
@@ -1,52 +1,11 @@
-"""
-Test using SDP-generated pools.
-"""
-
-import os
-import re
-from glob import glob
+"""Test using SDP-generated pools."""
 
 import pytest
 
-from jwst.associations.lib.diff import compare_asn_files
-from jwst.associations.main import Main as asn_generate
+from jwst.regtest.assoc_rt_helpers import assoc_sdp_against_standard, parfunc
 
 # Mark all tests in this module
 pytestmark = [pytest.mark.bigdata]
-
-# Decompose pool name to retrieve proposal and version id.
-pool_regex = re.compile(r"(?P<proposal>jw.+?)_(?P<versionid>.+)_pool")
-
-
-def parfunc(x):
-    """
-    For readable param ids.
-    ``x`` is tuple of ``(pool, args)``.
-    """
-    return x[0]
-
-
-def _assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args):
-    pool, args = pool_args
-    proposal, version_id = pool_regex.match(pool).group("proposal", "versionid")
-
-    input_csv = rtdata.get_data(f"associations/sdp/pools/{pool}.csv")
-    rtdata.output = os.curdir  # This test is jailed and parametrized
-    rtdata.okify_op = "sdp_pool_copy"  # failures as folder content replacements
-    # Create the associations
-    with resource_tracker.track(log=request):
-        asn_generate.cli(args + ["-p", rtdata.output, "--version-id", version_id, input_csv])
-    out_paths = sorted(glob("*.json"))
-
-    # Compare to the truth associations.
-    truth_pool_path = f"truth/test_associations_sdp_pools/{pool}/"
-    rtdata.truth_remote = truth_pool_path
-    truth_paths = sorted(
-        [rtdata.get_truth(p) for p in rtdata.data_glob(truth_pool_path, glob="*.json")]
-    )
-    if truth_paths == []:  # truth dir does not exist
-        rtdata.truth_remote = f"{rtdata._inputs_root}/{rtdata.env}/{truth_pool_path}"
-    compare_asn_files(out_paths, truth_paths)
 
 
 # NOTE: These are inflight equivalent approximate replacements for
@@ -99,50 +58,9 @@ def _assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args):
         ),
         pytest.param(("jw05204_20250308t202944_pool", []), id="pool_002_image_miri"),
     ],
-    ids=parfunc,
 )
 def test_std(_jail, rtdata, resource_tracker, request, pool_args):
-    _assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
-
-
-# NOTE: These are inflight equivalent approximate replacements for
-#       test_associations_standards.py test module (slow).
-@pytest.mark.parametrize(
-    "pool_args",
-    [
-        pytest.param(("jw01523_20250321t155408_pool", []), id="pool_009_spec_miri_lv2bkg_MRS"),
-        pytest.param(("jw01928_20250318t105523_pool", []), id="imprint_single_obs_1"),
-        pytest.param(
-            ("jw02084_20250320t081615_pool", []), id="pool_021_tso_NIRCAM_TSIMAGE_TSGRISM"
-        ),
-        pytest.param(("jw03383_20250307t235057_pool", []), id="pool_034_wfss_parallel_NIRISS"),
-        pytest.param(
-            ("jw03522_20250318t192624_pool", []), id="pool_024_nirspec_fss_nods_AND_SUBPXPTS4"
-        ),
-        pytest.param(("jw03702_20250320t002824_pool", []), id="imprint_2n_dithers_obs_1_2"),
-        pytest.param(("jw05221_20250411t182730_pool", []), id="pool_032_nircam_wfss"),
-        pytest.param(("jw05308_20250319t215735_pool", []), id="pool_027_nirspec_ifu_nods_4NODS"),
-        pytest.param(
-            ("jw05398_20250312t000159_default_pool", ["-i", "o041", "o042", "o043", "c1016"]),
-            id="pool_034_wfss_parallel_NIRCAM_DEFAULT",
-        ),  # This is "default" scenario: row grism, column grism, and direct image.
-        pytest.param(
-            (
-                "jw05398_20250312t000159_3r2d_pool",
-                ["-i", "o030", "o031", "o032", "o033", "o035", "c1012", "c1013", "c1014"],
-            ),
-            id="pool_034_wfss_parallel_NIRCAM_3ROW_2DIRECT",
-        ),  # Three row grism images and two direct images.
-        pytest.param(
-            ("jw05398_20250312t000159_row_pool", ["-i", "o036", "o037", "c1029"]),
-            id="pool_034_wfss_parallel_NIRCAM_ROW_ONLY",
-        ),  # Only row grism and direct image, but no column grism image.
-    ],
-    ids=parfunc,
-)
-@pytest.mark.slow
-def test_sslow(_jail, rtdata, resource_tracker, request, pool_args):
-    _assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
+    assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
 
 
 @pytest.mark.parametrize(
@@ -178,30 +96,4 @@ def test_sslow(_jail, rtdata, resource_tracker, request, pool_args):
     ids=parfunc,
 )
 def test_sdp(tmp_cwd, rtdata, resource_tracker, request, pool_args):
-    _assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
-
-
-@pytest.mark.parametrize(
-    "pool_args",
-    [
-        ("jw01192_o008_pool", ["--DMS", "-i", "o008"]),  # This pool checks imprint behavior.
-        ("jw01257_20221201t192226_pool", []),
-        ("jw01288_c1005_mostilno12_pool", ["-i", "o003", "c1001", "c1005"]),  # JP-3230
-        ("jw01290_20230304t140931_pool", []),
-        ("jw01480_20250319t173819_pool", []),  # NRC_GRISM, NRC_TSIMAGE, NRC_TSGRISM
-        ("jw01493_20230307t040130_pool", []),
-        ("jw02064_20230302t112350_pool", []),
-        (
-            "jw02739_20230710t150016_pool",
-            ["-i", "c1000"],
-        ),  # This association tests the Asn_Lv3ImageMosaic rule
-        ("jw02739_20230710t150016_o010_pool", ["-i", "o010"]),  # NRC_IMAGE
-        ("jw03855_20241103t042455_pool", []),
-        ("jw04237_20250321t192812_pool", []),  # Lvl 3 MIRI MRS BKG
-        ("jw04237_20250321t192812_dms_pool", ["--DMS"]),  # Lvl 3 MIRI MRS BKG with DMS
-    ],
-    ids=parfunc,
-)
-@pytest.mark.slow
-def test_slow(tmp_cwd, rtdata, resource_tracker, request, pool_args):
-    _assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
+    assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
